pyproject.toml
[tool.poetry]
name = "fastchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fastchain", from = "src"}]

[tool.poetry.dependencies]
python = "3.11.6"
fastapi = {extras = ["all"], version = "^0.104.1"}
langchain = "^0.0.330"
sqlalchemy = "^2.0.23"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
torch = "^2.1.1"
pypdf = "^3.17.1"
pandas = "^2.1.3"
tiktoken = "^0.5.1"
scikit-learn = "^1.3.2"
transformers = "^4.35.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
